more ops

+++ types.go
@@ -3,7 +3,7 @@ package main
 type O interface{}  // O represents any kind of value.
 type B = bool       // B represents booleans (0 and 1 but less memory)
 type F = float64    // F represents real numbers.
-type I = int64      // I represents integers.
+type I = int        // I represents integers.
 type S = string     // S represents (immutable) strings of bytes.
 type E error        // Errors (TODO: think about it)
 type M func(O) O    // M represents monadic functions
